requirements.txt
alembic==1.7.4
aniso8601==9.0.1
attrs==21.2.0
click==8.0.1
dnspython==2.1.0
dominate==2.6.0
email-validator==1.1.3
Flask==2.0.2
Flask-Admin==1.5.8
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.8.1
importlib-resources==5.2.2
itsdangerous==2.0.1
Jinja2==3.0.2
jsonschema==4.1.2
Mako==1.1.5
MarkupSafe==2.0.1
marshmallow==3.14.0
marshmallow-sqlalchemy==0.26.1
mistune==0.8.4
Pillow==8.3.2
psycopg2==2.9.1
pyrsistent==0.18.0
python-dotenv==0.19.0
python-slugify==5.0.2
pytz==2021.3
PyYAML==6.0
simplejson==3.17.5
six==1.16.0
SQLAlchemy==1.4.25
SQLAlchemy-serializer==1.4.1
text-unidecode==1.3
visitor==0.1.3
Werkzeug==2.0.2
WTForms==2.3.3
zipp==3.6.0
